+++ rs/embedders/src/wasmtime_embedder/signal_handler.rs
@@ -3,19 +3,20 @@ use crate::wasmtime_embedder::host_memory::MemoryPageSize;
 
 use ic_types::NumBytes;
 use libc::c_void;
-use memory_tracker::{SigsegvMemoryTracker, signal_access_kind_and_address};
+use memory_tracker::{
+    SigsegvMemoryTracker, signal_access_kind_and_address, signal_mutex::SignalMutex,
+};
 use std::convert::TryFrom;
-use std::sync::MutexGuard;
-use std::sync::{Arc, Mutex, atomic::Ordering};
+use std::sync::{Arc, atomic::Ordering};
 
 /// Helper function to create a memory tracking SIGSEGV handler function.
 pub(crate) fn sigsegv_memory_tracker_handler(
-    memories: Vec<(Arc<Mutex<SigsegvMemoryTracker>>, MemoryPageSize)>,
+    memories: Vec<(Arc<SignalMutex<SigsegvMemoryTracker>>, MemoryPageSize)>,
 ) -> impl Fn(i32, *const libc::siginfo_t, *const libc::c_void) -> bool + Send + Sync {
     let mut memories: Vec<_> = memories
         .into_iter()
         .map(|(t, size)| {
-            let base = t.lock().unwrap().memory_area().start();
+            let base = t.lock().memory_area().start();
             (base, t, size)
         })
         .collect();
@@ -23,7 +24,10 @@ pub(crate) fn sigsegv_memory_tracker_handler(
     memories.sort_by_key(|(base, _, _)| *base);
 
     let check_if_expanded =
-        move |tracker: &MutexGuard<SigsegvMemoryTracker>,
+        move |tracker: &memory_tracker::signal_mutex::SignalMutexGuard<
+            '_,
+            SigsegvMemoryTracker,
+        >,
               si_addr: *mut c_void,
               current_size_in_pages: &MemoryPageSize| unsafe {
             let page_count = current_size_in_pages.load(Ordering::SeqCst);
@@ -74,7 +78,7 @@ pub(crate) fn sigsegv_memory_tracker_handler(
             .find(|(base, _, _)| *base as *mut c_void <= si_addr)
             .unwrap_or(&memories[0]);
 
-        let memory_tracker = memory_tracker.lock().unwrap();
+        let memory_tracker = memory_tracker.lock();
         // Spawn a guard to report the total time spent in the handler.
         let _guard = scopeguard::guard(timer, |timer| {
             memory_tracker