+++ rs/memory_tracker/src/lib.rs
@@ -14,15 +14,18 @@ use std::{
     cell::Cell,
     ops::Range,
     sync::{
-        Arc, Mutex,
+        Arc,
         atomic::{AtomicU64, AtomicUsize, Ordering},
     },
     time::Duration,
 };
 
+use signal_mutex::SignalMutex;
+
 mod conversions;
 mod deterministic;
 mod prefetching;
+pub mod signal_mutex;
 #[cfg(test)]
 mod tests;
 
@@ -381,7 +384,7 @@ pub trait MemoryTracker {
         dirty_page_tracking: DirtyPageTracking,
         page_map: PageMap,
         memory_limits: MemoryLimits,
-        subtract_instruction_counter: Arc<Mutex<dyn FnMut(u64) + Send>>,
+        subtract_instruction_counter: Arc<SignalMutex<dyn FnMut(u64) + Send>>,
     ) -> nix::Result<Self>
     where
         Self: Sized;
@@ -437,7 +440,7 @@ pub fn new(
     page_map: PageMap,
     missing_page_handler_kind: Option<MissingPageHandlerKind>,
     memory_limits: MemoryLimits,
-    subtract_instruction_counter: Arc<Mutex<dyn FnMut(u64) + Send>>,
+    subtract_instruction_counter: Arc<SignalMutex<dyn FnMut(u64) + Send>>,
 ) -> nix::Result<SigsegvMemoryTracker> {
     match missing_page_handler_kind {
         Some(MissingPageHandlerKind::Deterministic) => {